requirements.txt
django-cors-headers
Django
# REST framework and JWT
djangorestframework
djangorestframework-simplejwt
openpyxl
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-pptx
pyspark>=3.3.0
duckdb>=0.8.1
numpy
pyarrow
tqdm
multiprocessing-logging
matplotlib
seaborn
plotly
scipy
scikit-learn
statsmodels
python-dotenv>=1.0.0
findspark
django-extensions
ydata-profiling>=4.0.0
django-sslserver-v2